+++ RayTracer.MathLibrary.Tests/Vector3Tests.cs
@@ -10,9 +10,9 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(1f, -1f, 0f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(1f, 1f, 6f, 3f, -2f, 5f, -2f, 3f, 1f)]
         [TestCase(0f, -4f, 0f, 0f, 0f, 0f, 0f, -4f, 0f)]
-        public void AdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                               float arg1X, float arg1Y, float arg1Z,
-                                                                                               float arg2X, float arg2Y, float arg2Z)
+        public void AdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                               double arg1X, double arg1Y, double arg1Z,
+                                                                                               double arg2X, double arg2Y, double arg2Z)
         {
             Vector3 vector1 = new Vector3()
             {
@@ -42,9 +42,9 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(2f, -2f, -1f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(-1f, 4f, 7f, 3f, -2f, 5f, -2f, 3f, 1f)]
         [TestCase(0f, -8f, 0f, 0f, 0f, 0f, 0f, -4f, 0f)]
-        public void DoubleAdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                                     float arg1X, float arg1Y, float arg1Z,
-                                                                                                     float arg2X, float arg2Y, float arg2Z)
+        public void DoubleAdditionOperator_WhenCalledWithTwoPoints_ReturnPointWithCombinedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                                     double arg1X, double arg1Y, double arg1Z,
+                                                                                                     double arg2X, double arg2Y, double arg2Z)
         {
             Vector3 vector1 = new Vector3()
             {
@@ -74,9 +74,9 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(-1f, 1f, 2f, 0f, 0f, 1f, 1f, -1f, -1f)]
         [TestCase(-2f, -4f, -6f, 3f, 2f, 1f, 5f, 6f, 7f)]
         [TestCase(0f, -4f, 0f, 0f, 0f, 0f, 0f, 4f, 0f)]
-        public void SubtractionOperator_WhenCalledWithTwoPoints_ReturnPointWithSubtractedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                                    float arg1X, float arg1Y, float arg1Z,
-                                                                                                    float arg2X, float arg2Y, float arg2Z)
+        public void SubtractionOperator_WhenCalledWithTwoPoints_ReturnPointWithSubtractedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                                    double arg1X, double arg1Y, double arg1Z,
+                                                                                                    double arg2X, double arg2Y, double arg2Z)
         {
             Vector3 vector1 = new Vector3()
             {
@@ -107,8 +107,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(3f, 3f, 3f, 2f, 2f, 2f, 1.5f)]
         [TestCase(3.5f, -7f, 10.5f, 1f, -2f, 3f, 3.5f)]
         [TestCase(0.5f, -1f, 1.5f, 1f, -2f, 3f, 0.5f)]
-        public void MultiplyOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                             float argX, float argY, float argZ, float scalar)
+        public void MultiplyOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                             double argX, double argY, double argZ, double scalar)
         {
             Vector3 expected = new Vector3()
             {
@@ -131,8 +131,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(1f, 1f, 1f, 2f, 2f, 2f, 2f)]
         [TestCase(2f, 2f, 2f, 3f, 3f, 3f, 1.5f)]
         [TestCase(0.5f, -1f, 1.5f, 1f, -2f, 3f, 2f)]
-        public void DivideOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                           float argX, float argY, float argZ, float scalar)
+        public void DivideOperator_WhenCalledWithPointAndScalar_ReturnMultipliedCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                           double argX, double argY, double argZ, double scalar)
         {
             Vector3 expected = new Vector3()
             {
@@ -155,8 +155,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(1f, 1f, 1f, -1f, -1f, -1f)]
         [TestCase(-2f, -2f, -2f, 2f, 2f, 2f)]
         [TestCase(1f, -2f, 3f, -1f, 2f, -3f)]
-        public void NegateVector_WhenCalledWithVector_ReturnVectorWithOppositeCordinates(float expectedX, float expectedY, float expectedZ,
-                                                                                         float argX, float argY, float argZ)
+        public void NegateVector_WhenCalledWithVector_ReturnVectorWithOppositeCordinates(double expectedX, double expectedY, double expectedZ,
+                                                                                         double argX, double argY, double argZ)
         {
             Vector3 expected = new Vector3()
             {
@@ -185,7 +185,7 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(5f, 4f, 0f, 3f, 0f)]
         [TestCase(3.74165f, -1f, -2f, -3f, 0.00001f)]
         [TestCase(3.74165f, 1f, 2f, 3f, 0.00001f)]
-        public void Magnitude_WhenCalled_ReturnLenghtOfVector(float expected, float argX, float argY, float argZ, float epsilon)
+        public void Magnitude_WhenCalled_ReturnLenghtOfVector(double expected, double argX, double argY, double argZ, double epsilon)
         {
             Vector3 vector = new Vector3()
             {
@@ -202,8 +202,8 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(0f, 0.707f, 0.707f, 0f, 2f, 2f, 0.001f)]
         [TestCase(1f, 0f, 0f, 4f, 0f, 0f, 0.00001f)]
         [TestCase(0.26726f, 0.53452f, 0.80178f, 1f, 2f, 3f, 0.00001f)]
-        public void Normalize_WhenCalledWithVector_ReturnNormalizedVector(float expectedX, float expectedY, float expectedZ,
-                                                                          float argX, float argY, float argZ, float epsilon)
+        public void Normalize_WhenCalledWithVector_ReturnNormalizedVector(double expectedX, double expectedY, double expectedZ,
+                                                                          double argX, double argY, double argZ, double epsilon)
         {
             Vector3 expected = new Vector3()
             {
@@ -227,7 +227,7 @@ namespace RayTracer.MathLibrary.Tests
         [TestCase(0f, 2f, 2f, 0.001f)]
         [TestCase(4f, 0f, 0f, 0f)]
         [TestCase(1f, 2f, 3f, 0.00001f)]
-        public void Normalize_WhenCalledWithVector_ChecksIfItsMagnitudeIs1(float argX, float argY, float argZ, float epsilon)
+        public void Normalize_WhenCalledWithVector_ChecksIfItsMagnitudeIs1(double argX, double argY, double argZ, double epsilon)
         {
             Vector3 vector = new Vector3()
             {
@@ -242,9 +242,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(20f, 1f, 2f, 3f, 2f, 3f, 4f)]
         [TestCase(20f, 2f, 3f, 4f, 1f, 2f, 3f)]
-        public void Dot_WhenCalledWithTwoVectors_ReturnDotProduct(float expected,
-                                                                  float arg1X, float arg1Y, float arg1Z,
-                                                                  float arg2X, float arg2Y, float arg2Z)
+        public void Dot_WhenCalledWithTwoVectors_ReturnDotProduct(double expected,
+                                                                  double arg1X, double arg1Y, double arg1Z,
+                                                                  double arg2X, double arg2Y, double arg2Z)
         {
             Vector3 vector1 = new Vector3()
             {
@@ -266,9 +266,9 @@ namespace RayTracer.MathLibrary.Tests
         [Test]
         [TestCase(-1f, 2f, -1f, 1f, 2f, 3f, 2f, 3f, 4f)]
         [TestCase(1f, -2f, 1f, 2f, 3f, 4f, 1f, 2f, 3f)]
-        public void Cross_WhenCalledWithTwoVectors_ReturnCrossProduct(float expectedX, float expectedY, float expectedZ,
-                                                                      float arg1X, float arg1Y, float arg1Z,
-                                                                      float arg2X, float arg2Y, float arg2Z)
+        public void Cross_WhenCalledWithTwoVectors_ReturnCrossProduct(double expectedX, double expectedY, double expectedZ,
+                                                                      double arg1X, double arg1Y, double arg1Z,
+                                                                      double arg2X, double arg2Y, double arg2Z)
         {
             Vector3 vector1 = new Vector3()
             {
