+++ RayTracer.MathLibrary/IColor.cs
@@ -6,10 +6,10 @@ namespace RayTracer.MathLibrary
 {
     public interface IColor
     {
-        public float Red { get; set; }
+        public double Red { get; set; }
 
-        public float Green { get; set; }
+        public double Green { get; set; }
 
-        public float Blue { get; set; }
+        public double Blue { get; set; }
     }
 }
